requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
apturl==0.5.2
astor==0.8.1
async-timeout==4.0.3
attrs==21.2.0
bcrypt==4.0.1
bleach==6.1.0
blessed==1.20.0
blinker==1.4
Brlapi==0.8.3
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==41.0.4
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.4.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.2
fasteners==0.14.1
filelock==3.12.3
frozenlist==1.4.0
fsspec==2023.9.0
future==0.18.2
git-python==1.0.3
gitdb==4.0.10
GitPython==3.1.36
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httplib2==0.20.2
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
inquirer==3.1.3
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
kaggle==1.5.16
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
litellm==0.1.609
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
open-interpreter==0.1.3
openai==0.27.10
orjson==3.9.7
packaging==23.1
paramiko==2.9.3
passlib==1.7.4
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pyasn1==0.5.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.6
python-nmap==0.7.1
python-slugify==8.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.1
readchar==4.0.5
regex==2023.8.8
reportlab==3.6.8
requests==2.31.0
rich==13.5.2
rsa==4.9
scapy==2.5.0
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
systemd-python==234
text-unidecode==1.3
texttable==1.6.4
tiktoken==0.4.0
tokentrim==0.1.10
tqdm==4.66.1
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.8.0
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.23.2
uvloop==0.17.0
wadllib==1.3.6
watchfiles==0.20.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.2.3
websockets==11.0.3
wget==3.2
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
